Make dbxtool.service not run on live images
If an OS is temporarily being booted on a machine, such as may be the
case with a Live Image, it's undesirable to have that OS make permanent
changes to the system's configuration and policy.  In the case of
dbxtool, that kind of security policy change could potentially include
the de-authorization of the installed OS, the developers of which are
expecting to apply such a change during the regular OS updates.  This
results in a *horrible* user experience, and we shouldn't encourage it.

At the same time, many OS installations start with by booting a Live
Image and copying it to become the installed system's root file system,
and in such cases we still want dbxtool to be installed and for the
service to run during the next reboot.  In that case we still need
dbxtool to be installed in the live image, just not to run.

This patch adds a criterion to the systemd service's [unit] declaration
to inhibit running if the kernel command line option "rd.live.image"
root "boot=live" are set, which should avoid this on both Fedora-like
and Debian-like systems.