Only keep the before last DBX revocation list file
We want to be more conservative in how the DBX updates are applied since a
vendor applying a DBX update to aggressively could cause another OS to not
boot anymore due not having a newer shim signed and all of them revoked.

To prevent this scenario, don't ship the latest DBX revocation file list
and only ship the before last. This means that the aa64 and ia32 DBX files
are removed since were only released starting with the latest update.

Also, each DBX revocation list released in https://uefi.org/ contains the
entries of all previous revocations. There's no need to keep the old ones.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>